♻️ refactor: Chart.tsx 코드 분리 #241 커스텀훅을 이용해 분리.
