+++ FE/src/components/Login/index.tsx
@@ -3,7 +3,7 @@ import Input from './Input';
 import { ChatBubbleOvalLeftIcon, XMarkIcon } from '@heroicons/react/16/solid';
 import { FormEvent, useEffect, useState } from 'react';
 import { login } from 'service/auth';
-import useAuthStore from 'store/authStore';
+import useAuthStore from 'store/useAuthStore.ts';
 import Overay from '../ModalOveray.tsx';
 
 export default function Login() {
@@ -59,7 +59,7 @@ export default function Login() {
       <Overay onClick={() => toggleModal()} />
       <section className='fixed left-1/2 top-1/2 flex w-[500px] -translate-x-1/2 -translate-y-1/2 flex-col rounded-2xl bg-white p-20 shadow-lg'>
         <h2 className='text-3xl font-bold'>JuGa</h2>
-        <p className='h-5 my-3 text-sm font-semibold text-juga-red-60'>
+        <p className='my-3 h-5 text-sm font-semibold text-juga-red-60'>
           {
             {
               '401': '존재하지 않는 사용자입니다.',
@@ -67,8 +67,8 @@ export default function Login() {
             }[errorCode]
           }
         </p>
-        <form className='flex flex-col mb-2' onSubmit={handleSubmit}>
-          <div className='flex flex-col gap-2 my-10'>
+        <form className='mb-2 flex flex-col' onSubmit={handleSubmit}>
+          <div className='my-10 flex flex-col gap-2'>
             <Input
               type='text'
               placeholder='아이디'
@@ -84,7 +84,7 @@ export default function Login() {
               autoComplete='current-password'
             />
           </div>
-          <button className='py-2 text-white transition rounded-3xl bg-juga-blue-40 hover:bg-juga-blue-50'>
+          <button className='rounded-3xl bg-juga-blue-40 py-2 text-white transition hover:bg-juga-blue-50'>
             로그인
           </button>
         </form>
