+++ FE/src/components/StocksDetail/Chart.tsx
@@ -1,12 +1,5 @@
-import {
-  MouseEvent,
-  useCallback,
-  useEffect,
-  useRef,
-  useState,
-  WheelEvent,
-} from 'react';
-import { Padding, StockChartUnit, TiemCategory } from 'types';
+import { useCallback, useEffect, useRef, useState } from 'react';
+import { MousePositionType, StockChartUnit, TiemCategory } from 'types';
 import { useQuery } from '@tanstack/react-query';
 import { getStocksChartDataByCode } from 'service/stocks';
 import { drawLineChart } from 'utils/chart/drawLineChart.ts';
@@ -18,61 +11,31 @@ import { drawLowerYAxis } from 'utils/chart/drawLowerYAxis.ts';
 import { drawChartGrid } from 'utils/chart/drawChartGrid.ts';
 import { drawMouseGrid } from 'utils/chart/drawMouseGrid.ts';
 import { EyeIcon, EyeSlashIcon } from '@heroicons/react/16/solid';
-import { ChartSizeConfigType } from './type.ts';
+import { setCanvasSize } from 'utils/chart/setCanvasSize.ts';
+import { useMouseMove } from 'hooks/useMouseMove.ts';
+import { useMouseUpDown } from 'hooks/useMouseUpDown.ts';
+import { useMouseWheel } from 'hooks/useMouseWheel.ts';
+import { categories, padding } from 'constants.ts';
+import { useCanvasRef } from 'hooks/useCanvasRef.ts';
+import { useCanvasResize } from 'hooks/useCanvasResize.ts';
 
-const categories: { label: string; value: TiemCategory }[] = [
-  { label: '일', value: 'D' },
-  { label: '주', value: 'W' },
-  { label: '월', value: 'M' },
-  { label: '년', value: 'Y' },
-];
-
-const padding: Padding = {
-  top: 20,
-  right: 80,
-  bottom: 10,
-  left: 40,
-};
-
-type StocksDeatailChartProps = {
+type StocksDetailChartProps = {
   code: string;
 };
 
-export type MousePositionType = {
-  x: number;
-  y: number;
-};
-
-export default function Chart({ code }: StocksDeatailChartProps) {
+export default function Chart({ code }: StocksDetailChartProps) {
   const containerRef = useRef<HTMLDivElement>(null);
-  const upperChartCanvasRef = useRef<HTMLCanvasElement>(null);
-  const lowerChartCanvasRef = useRef<HTMLCanvasElement>(null);
-  const upperChartY = useRef<HTMLCanvasElement>(null);
-  const lowerChartY = useRef<HTMLCanvasElement>(null);
-  const chartX = useRef<HTMLCanvasElement>(null);
-  const rafRef = useRef<number>();
+  const {
+    upperChartCanvasRef,
+    lowerChartCanvasRef,
+    upperChartY,
+    lowerChartY,
+    chartX,
+  } = useCanvasRef();
   const [timeCategory, setTimeCategory] = useState<TiemCategory>('D');
   const [moveAverageToggle, setMoveAverageToggle] = useState(true);
-  const [charSizeConfig, setChartSizeConfig] = useState<ChartSizeConfigType>({
-    upperHeight: 0.5,
-    lowerHeight: 0.4,
-    chartWidth: 0.92,
-    yAxisWidth: 0.08,
-    xAxisHeight: 0.1,
-  });
-  const [isDragging, setIsDragging] = useState(false);
-  const [upperLabelNum, setUpperLabelNum] = useState(3);
-  const [lowerLabelNum, setLowerLabelNum] = useState(3);
-  const [mousePosition, setMousePosition] = useState<MousePositionType>({
-    x: 0,
-    y: 0,
-  });
+
   const [mouseIndex, setMouseIndex] = useState<number | null>(null);
-  const [dataRange, setDataRange] = useState({
-    start: 0,
-    end: 0,
-  });
-  const minDisplayData = 20;
 
   const { data, isLoading } = useQuery(
     ['stocksChartData', code, timeCategory],
@@ -80,95 +43,20 @@ export default function Chart({ code }: StocksDeatailChartProps) {
     { staleTime: 1000 * 60 },
   );
 
-  const handleMouseDown = useCallback((e: MouseEvent) => {
-    e.preventDefault();
-    setIsDragging(true);
-  }, []);
-
-  const handleMouseMove = useCallback(
-    (e: globalThis.MouseEvent) => {
-      if (!isDragging || !containerRef.current) return;
-      const minHeight = 0.2;
-      const containerRect = containerRef.current.getBoundingClientRect();
-      const mouseY = e.clientY - containerRect.top;
-      const ratio = mouseY / containerRef.current.clientHeight;
-      const maxHeight = 0.9 - minHeight;
-      const upperRatio = Math.min(maxHeight, Math.max(minHeight, ratio));
-      const lowerRatio = 0.9 - upperRatio;
-
-      const calculateLabelNum = (ratio: number) => {
-        if (ratio <= 0.2) return 1;
-        if (ratio <= 0.35) return 2;
-        if (ratio <= 0.55) return 3;
-        return 4;
-      };
-
-      if (lowerRatio >= minHeight && upperRatio >= minHeight) {
-        setChartSizeConfig((prev) => ({
-          ...prev,
-          upperHeight: upperRatio,
-          lowerHeight: lowerRatio,
-        }));
-
-        setUpperLabelNum(calculateLabelNum(upperRatio));
-        setLowerLabelNum(calculateLabelNum(lowerRatio));
-      }
+  const { mousePosition, getCanvasMousePosition } = useMouseMove(containerRef);
+  const { chartSizeConfig, upperLabelNum, lowerLabelNum, handleMouseDown } =
+    useMouseUpDown(containerRef);
+  const { dataRange, handleWheel } = useMouseWheel(data);
+  const { resizeCanvases } = useCanvasResize(
+    {
+      upperChartCanvasRef,
+      lowerChartCanvasRef,
+      upperChartY,
+      lowerChartY,
+      chartX,
     },
-    [
-      isDragging,
-      containerRef,
-      setChartSizeConfig,
-      setUpperLabelNum,
-      setLowerLabelNum,
-    ],
-  );
-
-  const handleMouseUp = useCallback(() => {
-    setIsDragging(false);
-  }, []);
-
-  const getCanvasMousePosition = (e: MouseEvent) => {
-    if (!containerRef.current) return;
-
-    if (rafRef.current) {
-      cancelAnimationFrame(rafRef.current);
-    }
-
-    rafRef.current = requestAnimationFrame(() => {
-      const rect = containerRef.current?.getBoundingClientRect();
-      if (!rect) return;
-
-      setMousePosition({
-        x: (e.clientX - rect.left) * 2,
-        y: (e.clientY - rect.top) * 2,
-      });
-    });
-  };
-
-  useEffect(() => {
-    if (isDragging) {
-      window.addEventListener('mousemove', handleMouseMove);
-      window.addEventListener('mouseup', handleMouseUp);
-    }
-
-    return () => {
-      window.removeEventListener('mousemove', handleMouseMove);
-      window.removeEventListener('mouseup', handleMouseUp);
-      if (rafRef.current) {
-        cancelAnimationFrame(rafRef.current);
-      }
-    };
-  }, [isDragging, handleMouseDown, handleMouseUp, handleMouseMove]);
-  const setCanvasSize = useCallback(
-    (canvas: HTMLCanvasElement, widthConfig: number, heightConfig: number) => {
-      if (!containerRef.current) return;
-
-      canvas.width = containerRef.current.clientWidth * widthConfig * 2;
-      canvas.height = containerRef.current.clientHeight * heightConfig * 2;
-      canvas.style.width = `${containerRef.current.clientWidth * widthConfig}px`;
-      canvas.style.height = `${containerRef.current.clientHeight * heightConfig}px`;
-    },
-    [containerRef],
+    chartSizeConfig,
+    containerRef,
   );
 
   const renderChart = useCallback(
@@ -311,49 +199,6 @@ export default function Chart({ code }: StocksDeatailChartProps) {
       dataRange,
     ],
   );
-  const handleWheel = useCallback(
-    (e: WheelEvent<HTMLDivElement>) => {
-      if (!data) return;
-
-      const wheelPower = 0.1;
-      const curRange = dataRange.end - dataRange.start;
-
-      // 축소
-      if (e.deltaY > 0) {
-        // 늘어나야함 & 데이터 최대 갯수보다 작아야함.
-        const newRange = Math.min(curRange * (1 + wheelPower), data.length);
-        const newStart = Math.max(data.length - newRange, 0);
-
-        setDataRange({
-          start: newStart,
-          end: data.length - 1,
-        });
-      }
-
-      // 확대
-      if (e.deltaY < 0) {
-        // 줄어야함 & 최소 데이터 갯수보단 커야함.
-        const newRange = Math.max(curRange * (1 - wheelPower), minDisplayData);
-        const newStart = Math.max(data.length - newRange, 0);
-
-        setDataRange({
-          start: newStart,
-          end: data.length - 1,
-        });
-      }
-    },
-    [data, dataRange],
-  );
-
-  useEffect(() => {
-    if (data) {
-      setDataRange((prev) => ({
-        ...prev,
-        start: 0,
-        end: data.length - 1,
-      }));
-    }
-  }, [data]);
 
   useEffect(() => {
     if (isLoading || !data) return;
@@ -366,35 +211,8 @@ export default function Chart({ code }: StocksDeatailChartProps) {
       !chartX.current
     )
       return;
-    setCanvasSize(
-      upperChartCanvasRef.current,
-      charSizeConfig.chartWidth,
-      charSizeConfig.upperHeight,
-    );
 
-    setCanvasSize(
-      upperChartY.current,
-      charSizeConfig.yAxisWidth,
-      charSizeConfig.upperHeight,
-    );
-
-    setCanvasSize(
-      lowerChartCanvasRef.current,
-      charSizeConfig.chartWidth,
-      charSizeConfig.lowerHeight,
-    );
-
-    setCanvasSize(
-      lowerChartY.current,
-      charSizeConfig.yAxisWidth,
-      charSizeConfig.lowerHeight,
-    );
-
-    setCanvasSize(
-      chartX.current,
-      charSizeConfig.chartWidth,
-      charSizeConfig.xAxisHeight,
-    );
+    resizeCanvases();
 
     renderChart(
       upperChartCanvasRef.current,
@@ -411,9 +229,10 @@ export default function Chart({ code }: StocksDeatailChartProps) {
     isLoading,
     setCanvasSize,
     renderChart,
-    charSizeConfig,
+    chartSizeConfig,
     mousePosition,
   ]);
+
   return (
     <div className='box-border flex h-[260px] flex-col items-center rounded-lg bg-white p-3'>
       <div className='flex h-fit w-full items-center justify-between'>
@@ -441,24 +260,6 @@ export default function Chart({ code }: StocksDeatailChartProps) {
           <canvas ref={upperChartCanvasRef} />
           <canvas ref={upperChartY} />
           <div className='absolute flex w-[520px] flex-col gap-1'>
-            {mouseIndex !== null && data ? (
-              <div
-                className={'relative flex flex-row items-center gap-1 text-xs'}
-              >
-                <span>
-                  시작 {Number(data[mouseIndex].stck_oprc).toLocaleString()}원
-                </span>
-                <span>
-                  고가 {Number(data[mouseIndex].stck_hgpr).toLocaleString()}원
-                </span>
-                <span>
-                  저가 {Number(data[mouseIndex].stck_lwpr).toLocaleString()}원
-                </span>
-                <span>
-                  종가 {Number(data[mouseIndex].stck_clpr).toLocaleString()}원
-                </span>
-              </div>
-            ) : null}
             <div className={'relative flex flex-row items-center gap-1'}>
               <button
                 className='flex h-3 w-3 items-center justify-center rounded bg-juga-grayscale-50'
@@ -505,6 +306,24 @@ export default function Chart({ code }: StocksDeatailChartProps) {
                 )}
               </div>
             </div>
+            {mouseIndex !== null && data ? (
+              <div
+                className={'relative flex flex-row items-center gap-1 text-xs'}
+              >
+                <span>
+                  시작 {Number(data[mouseIndex].stck_oprc).toLocaleString()}원
+                </span>
+                <span>
+                  고가 {Number(data[mouseIndex].stck_hgpr).toLocaleString()}원
+                </span>
+                <span>
+                  저가 {Number(data[mouseIndex].stck_lwpr).toLocaleString()}원
+                </span>
+                <span>
+                  종가 {Number(data[mouseIndex].stck_clpr).toLocaleString()}원
+                </span>
+              </div>
+            ) : null}
           </div>
         </div>
         <div className='group flex h-[1px] w-full cursor-row-resize items-center justify-center bg-juga-grayscale-100'>
