+++ FE/src/hooks/useChart.ts
@@ -1,0 +1,11 @@
+import { useRef } from 'react';
+
+export const useChart = () => {
+  const upperChartCanvasRef = useRef<HTMLCanvasElement>(null);
+
+  // const upperChartY = useRef<HTMLCanvasElement>(null);
+
+  return {
+    upperChartCanvasRef,
+  };
+};
